refactor(frontend): Rename Ethereum WalletConnect send modal (#8927)
# Motivation

For consistency, `WalletConnectSendModal` becomes
`EthWalletConnectSendModal`.
