+++ src/frontend/src/eth/components/wallet-connect/EthWalletConnectSendModal.svelte
